feat(package): validation code for schema v1.2.1
Added the accessor overrides, the schema validator, entry point validation, and tools validation in order to enable the logic to validate a dual-entry schema.
The dual entry is both the mcp server code and the hatch wrapper code.
Also enforces to write all tools in the mcp_server.py to easily be compatible with other clients supporting MCP.