fix: access to entrypoints
The accessor of a package following v1.2.1 of the package metadata schema would return an object containing the two new entry point definitions.
But before it was returning a dict. Hence the consummers of the API of the validator would suddenly get broken because the return type was not the same anymore.
Following the pattern of increase coverage of fields via the chain of responsibility, we added virtual functions `get_mcp_entry_point` and `get_hatch_mcp_entry_point` to `pkg_accessor_base.py`
Then, v1.1.0 implements the default lower level which points to the same as the already existing `get_entry_point`.
The accessor for v1.2.0 delegates to v1.1.0.
The accessor for v1.2.1 returns the values associated to the keys `mcp_server` and `hatch_mcp_server` respectively. In addition, the legacy `get_entry_point` points to `mcp_server.py`